feat(pipeline): improve dashboard query sorting, adjust battery type, and integrate SSE port interface

+++ src/main/java/cloud/chlora/pipeline/dashboard/domain/model/DashboardSnapshot.java
@@ -24,7 +24,7 @@ public record DashboardSnapshot(
             String potName,
             boolean isOnline,
             float temperature,
-            float battery,
+            int battery,
             float soilMoisture,
             float humidity,
             int anomalyCount,
